:hammer: Centralize common dataset functionality; rename max_size_cache to max_cache_size in dataset config files;
